requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
apriltag==0.0.16
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.9
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
idna==3.7
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
kagglehub==0.3.3
keras==3.6.0
kiwisolver==1.4.7
lapx==0.5.11
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxmltools==1.12.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
proglog==0.1.10
protobuf==3.20.3
psutil==6.1.0
psycopg2==2.9.10
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.48
scipy==1.14.1
seaborn==0.13.2
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
supervision==0.24.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tf2onnx==1.16.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.26
ultralytics-thop==2.0.10
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.0
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.16.0
